backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
google-generativeai==0.8.3
pydantic==2.9.2
jinja2==3.1.4
aiosmtplib==3.0.2
aiofiles==24.1.0
